requirements.txt
python-dotenv==1.0.0
requests==2.31.0
openai==1.1.0
pytest==7.4.3
pydantic==2.5.0
fastapi==0.110.0
uvicorn==0.27.0
jinja2==3.1.2
python-multipart==0.0.9